# [android/ios] improve perf when not using Application.Properties