--- a/Xamarin.Forms.Platform.Android/Deserializer.cs
+++ b/Xamarin.Forms.Platform.Android/Deserializer.cs
@@ -57,16 +57,16 @@ namespace Xamarin.Forms.Platform.Android
 			{
 				using (IsolatedStorageFile store = IsolatedStorageFile.GetUserStoreForApplication())
 				{
+					// No need to write 0 properties if no file exists
+					if (properties.Count == 0 && !store.FileExists(PropertyStoreFile))
+					{
+						return;
+					}
 					using (IsolatedStorageFileStream stream = store.OpenFile(PropertyStoreFile + ".tmp", System.IO.FileMode.OpenOrCreate))
 					using (XmlDictionaryWriter writer = XmlDictionaryWriter.CreateBinaryWriter(stream))
 					{
 						try
 						{
-							// No need to write 0 properties if no file exists
-							if (properties.Count == 0 && !store.FileExists(PropertyStoreFile))
-							{
-								return;
-							}
 							var dcs = new DataContractSerializer(typeof(Dictionary<string, object>));
 							dcs.WriteObject(writer, properties);
 							writer.Flush();