--- a/Xamarin.Forms.Platform.Tizen/Deserializer.cs
+++ b/Xamarin.Forms.Platform.Tizen/Deserializer.cs
@@ -1,15 +1,14 @@
 using System;
 using System.Collections.Generic;
-using System.Threading.Tasks;
-using System.Runtime.Serialization;
-using System.Xml;
 using System.Diagnostics;
-using System.IO;
-using Xamarin.Forms.Internals;
+using System.IO.IsolatedStorage;
+using System.Runtime.Serialization;
+using System.Threading.Tasks;
+using System.Xml;
 
 namespace Xamarin.Forms.Platform.Tizen
 {
-	internal class Deserializer : IDeserializer
+	internal class Deserializer : Internals.IDeserializer
 	{
 		const string PropertyStoreFile = "PropertyStore.forms";
 
@@ -19,33 +18,29 @@ namespace Xamarin.Forms.Platform.Tizen
 			// Make sure to use Internal
 			return Task.Run(() =>
 			{
-				var store = new TizenIsolatedStorageFile();
-				Stream stream = null;
-				try
+				using (IsolatedStorageFile store = IsolatedStorageFile.GetUserStoreForApplication())
 				{
-					stream = store.OpenFile(PropertyStoreFile, System.IO.FileMode.OpenOrCreate);
-					if (stream.Length == 0)
-					{
+					if (!store.FileExists(PropertyStoreFile))
 						return null;
-					}
+
+					using (IsolatedStorageFileStream stream = store.OpenFile(PropertyStoreFile, System.IO.FileMode.Open, System.IO.FileAccess.Read))
 					using (XmlDictionaryReader reader = XmlDictionaryReader.CreateBinaryReader(stream, XmlDictionaryReaderQuotas.Max))
 					{
-						stream = null;
-						var dcs = new DataContractSerializer(typeof(Dictionary<string, object>));
-						return (IDictionary<string, object>)dcs.ReadObject(reader);
+						if (stream.Length == 0)
+							return null;
+
+						try
+						{
+							var dcs = new DataContractSerializer(typeof(Dictionary<string, object>));
+							return (IDictionary<string, object>)dcs.ReadObject(reader);
+						}
+						catch (Exception e)
+						{
+							Debug.WriteLine("Could not deserialize properties: " + e.Message);
+							Internals.Log.Warning("Xamarin.Forms PropertyStore", $"Exception while reading Application properties: {e}");
+						}
 					}
-				}
-				catch (Exception e)
-				{
-					Debug.WriteLine("Could not deserialize properties: " + e.Message);
-					Internals.Log.Warning("Xamarin.Forms PropertyStore", $"Exception while reading Application properties: {e}");
-				}
-				finally
-				{
-					if (stream != null)
-					{
-						stream.Dispose();
-					}
+
 				}
 
 				return null;
@@ -59,52 +54,41 @@ namespace Xamarin.Forms.Platform.Tizen
 			// Make sure to use Internal
 			return Task.Run(() =>
 			{
-				var success = false;
-				var store = new TizenIsolatedStorageFile();
-				Stream stream = null;
-				try
+				using (IsolatedStorageFile store = IsolatedStorageFile.GetUserStoreForApplication())
 				{
 					// No need to write 0 properties if no file exists
 					if (properties.Count == 0 && !store.FileExists(PropertyStoreFile))
 					{
 						return;
 					}
-					stream = store.OpenFile(PropertyStoreFile + ".tmp", System.IO.FileMode.OpenOrCreate);
+					using (IsolatedStorageFileStream stream = store.OpenFile(PropertyStoreFile + ".tmp", System.IO.FileMode.OpenOrCreate))
 					using (XmlDictionaryWriter writer = XmlDictionaryWriter.CreateBinaryWriter(stream))
 					{
-						stream = null;
-						var dcs = new DataContractSerializer(typeof(Dictionary<string, object>));
-						dcs.WriteObject(writer, properties);
-						writer.Flush();
-						success = true;
+						try
+						{
+							var dcs = new DataContractSerializer(typeof(Dictionary<string, object>));
+							dcs.WriteObject(writer, properties);
+							writer.Flush();
+						}
+						catch (Exception e)
+						{
+							Debug.WriteLine("Could not serialize properties: " + e.Message);
+							Internals.Log.Warning("Xamarin.Forms PropertyStore", $"Exception while writing Application properties: {e}");
+							return;
+						}
 					}
-				}
-				catch (Exception e)
-				{
-					Debug.WriteLine("Could not serialize properties: " + e.Message);
-					Internals.Log.Warning("Xamarin.Forms PropertyStore", $"Exception while writing Application properties: {e}");
-				}
-				finally
-				{
-					if (stream != null)
+
+					try
 					{
-						stream.Dispose();
+						if (store.FileExists(PropertyStoreFile))
+							store.DeleteFile(PropertyStoreFile);
+						store.MoveFile(PropertyStoreFile + ".tmp", PropertyStoreFile);
 					}
-				}
-
-				if (!success)
-					return;
-
-				try
-				{
-					if (store.FileExists(PropertyStoreFile))
-						store.DeleteFile(PropertyStoreFile);
-					store.MoveFile(PropertyStoreFile + ".tmp", PropertyStoreFile);
-				}
-				catch (Exception e)
-				{
-					Debug.WriteLine("Could not move new serialized property file over old: " + e.Message);
-					Internals.Log.Warning("Xamarin.Forms PropertyStore", $"Exception while writing Application properties: {e}");
+					catch (Exception e)
+					{
+						Debug.WriteLine("Could not move new serialized property file over old: " + e.Message);
+						Internals.Log.Warning("Xamarin.Forms PropertyStore", $"Exception while writing Application properties: {e}");
+					}
 				}
 			});
 		}