--- a/Xamarin.Forms.Platform.Tizen/Deserializer.cs
+++ b/Xamarin.Forms.Platform.Tizen/Deserializer.cs
@@ -64,6 +64,11 @@ namespace Xamarin.Forms.Platform.Tizen
 				Stream stream = null;
 				try
 				{
+					// No need to write 0 properties if no file exists
+					if (properties.Count == 0 && !store.FileExists(PropertyStoreFile))
+					{
+						return;
+					}
 					stream = store.OpenFile(PropertyStoreFile + ".tmp", System.IO.FileMode.OpenOrCreate);
 					using (XmlDictionaryWriter writer = XmlDictionaryWriter.CreateBinaryWriter(stream))
 					{