--- a/Xamarin.Forms.Platform.iOS/Deserializer.cs
+++ b/Xamarin.Forms.Platform.iOS/Deserializer.cs
@@ -54,29 +54,30 @@ namespace Xamarin.Forms.Platform.MacOS
 			// Make sure to use Internal
 			return Task.Run(() =>
 			{
-				var success = false;
 				using (var store = IsolatedStorageFile.GetUserStoreForApplication())
-				using (var stream = store.OpenFile(PropertyStoreFile + ".tmp", System.IO.FileMode.OpenOrCreate))
-				using (var writer = XmlDictionaryWriter.CreateBinaryWriter(stream))
 				{
-					try
+					using (var stream = store.OpenFile(PropertyStoreFile + ".tmp", System.IO.FileMode.OpenOrCreate))
+					using (var writer = XmlDictionaryWriter.CreateBinaryWriter(stream))
 					{
-						var dcs = new DataContractSerializer(typeof(Dictionary<string, object>));
-						dcs.WriteObject(writer, properties);
-						writer.Flush();
-						success = true;
+						try
+						{
+							// No need to write 0 properties if no file exists
+							if (properties.Count == 0 && !store.FileExists(PropertyStoreFile))
+							{
+								return;
+							}
+							var dcs = new DataContractSerializer(typeof(Dictionary<string, object>));
+							dcs.WriteObject(writer, properties);
+							writer.Flush();
+						}
+						catch (Exception e)
+						{
+							Debug.WriteLine("Could not serialize properties: " + e.Message);
+							Log.Warning("Xamarin.Forms PropertyStore", $"Exception while writing Application properties: {e}");
+							return;
+						}
 					}
-					catch (Exception e)
-					{
-						Debug.WriteLine("Could not serialize properties: " + e.Message);
-						Log.Warning("Xamarin.Forms PropertyStore", $"Exception while writing Application properties: {e}");
-					}
-				}
 
-				if (!success)
-					return;
-				using (var store = IsolatedStorageFile.GetUserStoreForApplication())
-				{
 					try
 					{
 						if (store.FileExists(PropertyStoreFile))