@jonathanpeppers 2020-01-07
> To know if a 10ms improvement is perceptible we'd have to know the total startup time in which the improvement was measured.

The log messages above have this?

The `ActivityTaskManager`/`ActivityManager` log message is a system process on Android. It logs the total display time for each activity the system displays. More on this here:

https://developer.android.com/topic/performance/vitals/launch-time#time-initial

> In Android 4.4 (API level 19) and higher, logcat includes an output line containing a value called Displayed. This value represents the amount of time elapsed between launching the process and finishing drawing the corresponding activity on the screen.

What are you asking for here? Thanks!